Create TalkApp.java
This is a chat app. It is not very impressive but it simply depicts the function of a chatting app. There is simple functionality involved in this app which allows you to interact with it in a fun manner. You can simply write anything in the message box and look for the answer. If your message doesn't match with it's any function, it will prompt a message that what can you ask from it. You can also add your own custom messages to make it more interactive.